+++ router.go
@@ -12,6 +12,10 @@ func (r *router) Get(path string, handler http.Handler) {
 	r.Route(http.MethodGet, path, handler)
 }
 
+func (r *router) Head(path string, handler http.Handler) {
+	r.Route(http.MethodHead, path, handler)
+}
+
 func (r *router) Post(path string, handler http.Handler) {
 	r.Route(http.MethodPost, path, handler)
 }
@@ -28,8 +32,16 @@ func (r *router) Delete(path string, handler http.Handler) {
 	r.Route(http.MethodDelete, path, handler)
 }
 
+func (r *router) Connect(path string, handler http.Handler) {
+	r.Route(http.MethodConnect, path, handler)
+}
+
 func (r *router) Options(path string, handler http.Handler) {
 	r.Route(http.MethodOptions, path, handler)
+}
+
+func (r *router) Trace(path string, handler http.Handler) {
+	r.Route(http.MethodTrace, path, handler)
 }
 
 func (r *router) Route(method string, path string, handler http.Handler) {