+++ router_method_test.go
@@ -16,7 +16,7 @@ func (h handler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	return
 }
 
-func TestRouter(t *testing.T) {
+func TestRouterMethod(t *testing.T) {
 	t.Run("GET ok", func(t *testing.T) {
 		req, _ := http.NewRequest("GET", "/", nil)
 		rr := httptest.NewRecorder()
@@ -37,6 +37,34 @@ func TestRouter(t *testing.T) {
 
 		router := New()
 		router.Get("/", newHandler())
+
+		router.Mux().ServeHTTP(rr, req)
+
+		if status := rr.Code; status != http.StatusMethodNotAllowed {
+			t.Errorf("Expecting status code %d, got %d instead", http.StatusMethodNotAllowed, status)
+		}
+	})
+
+	t.Run("HEAD ok", func(t *testing.T) {
+		req, _ := http.NewRequest("HEAD", "/", nil)
+		rr := httptest.NewRecorder()
+
+		router := New()
+		router.Head("/", newHandler())
+
+		router.Mux().ServeHTTP(rr, req)
+
+		if status := rr.Code; status != http.StatusOK {
+			t.Errorf("Expecting status code %d, got %d instead", http.StatusOK, status)
+		}
+	})
+
+	t.Run("HEAD not ok", func(t *testing.T) {
+		req, _ := http.NewRequest("HEAD", "/", nil)
+		rr := httptest.NewRecorder()
+
+		router := New()
+		router.Connect("/", newHandler())
 
 		router.Mux().ServeHTTP(rr, req)
 
@@ -157,6 +185,34 @@ func TestRouter(t *testing.T) {
 		}
 	})
 
+	t.Run("CONNECT ok", func(t *testing.T) {
+		req, _ := http.NewRequest("CONNECT", "/", nil)
+		rr := httptest.NewRecorder()
+
+		router := New()
+		router.Connect("/", newHandler())
+
+		router.Mux().ServeHTTP(rr, req)
+
+		if status := rr.Code; status != http.StatusOK {
+			t.Errorf("Expecting status code %d, got %d instead", http.StatusOK, status)
+		}
+	})
+
+	t.Run("CONNECT not ok", func(t *testing.T) {
+		req, _ := http.NewRequest("CONNECT", "/", nil)
+		rr := httptest.NewRecorder()
+
+		router := New()
+		router.Post("/", newHandler())
+
+		router.Mux().ServeHTTP(rr, req)
+
+		if status := rr.Code; status != http.StatusMethodNotAllowed {
+			t.Errorf("Expecting status code %d, got %d instead", http.StatusMethodNotAllowed, status)
+		}
+	})
+
 	t.Run("OPTIONS ok", func(t *testing.T) {
 		req, _ := http.NewRequest("OPTIONS", "/", nil)
 		rr := httptest.NewRecorder()
@@ -177,6 +233,34 @@ func TestRouter(t *testing.T) {
 
 		router := New()
 		router.Patch("/", newHandler())
+
+		router.Mux().ServeHTTP(rr, req)
+
+		if status := rr.Code; status != http.StatusMethodNotAllowed {
+			t.Errorf("Expecting status code %d, got %d instead", http.StatusMethodNotAllowed, status)
+		}
+	})
+
+	t.Run("TRACE ok", func(t *testing.T) {
+		req, _ := http.NewRequest("TRACE", "/", nil)
+		rr := httptest.NewRecorder()
+
+		router := New()
+		router.Trace("/", newHandler())
+
+		router.Mux().ServeHTTP(rr, req)
+
+		if status := rr.Code; status != http.StatusOK {
+			t.Errorf("Expecting status code %d, got %d instead", http.StatusOK, status)
+		}
+	})
+
+	t.Run("TRACE not ok", func(t *testing.T) {
+		req, _ := http.NewRequest("TRACE", "/", nil)
+		rr := httptest.NewRecorder()
+
+		router := New()
+		router.Get("/", newHandler())
 
 		router.Mux().ServeHTTP(rr, req)